rename file from src/cli/command.ts to src/cli/legacy-command.ts
