+++ src/cli/commands/private-cmds/migrate-cmd.ts
@@ -1,5 +1,5 @@
 import chalk from 'chalk';
-import { LegacyCommand, CommandOptions } from '../../command';
+import { LegacyCommand, CommandOptions } from '../../legacy-command';
 import { migrate } from '../../../api/consumer';
 
 export default class Migrate implements LegacyCommand {
