+++ src/cli/commands/private-cmds/resolver-cmd.ts
@@ -1,7 +1,7 @@
 /* eslint max-classes-per-file: 0 */
 
 import chalk from 'chalk';
-import { LegacyCommand } from '../../command';
+import { LegacyCommand } from '../../legacy-command';
 import { getResolver, setResolver, resetResolver } from '../../../api/scope/lib/resolver';
 
 class ResolverSet implements LegacyCommand {
