+++ src/cli/commands/public-cmds/clear-cache-cmd.ts
@@ -2,7 +2,7 @@ import chalk from 'chalk';
 import fs from 'fs-extra';
 // it's a hack, but I didn't find a better way to access the getCacheDir() function
 import { __TEST__ as v8CompileCache } from 'v8-compile-cache';
-import { LegacyCommand } from '../../command';
+import { LegacyCommand } from '../../legacy-command';
 
 const { BASE_DOCS_DOMAIN } = require('../../../constants');
