+++ src/cli/commands/public-cmds/config-cmd.ts
@@ -1,7 +1,7 @@
 /* eslint max-classes-per-file: 0 */
 import rightpad from 'pad-right';
 import chalk from 'chalk';
-import { LegacyCommand } from '../../command';
+import { LegacyCommand } from '../../legacy-command';
 import { BASE_DOCS_DOMAIN } from '../../../constants';
 // import { config } from '../../../api/consumer';
 // const config = require('../../../api/consumer/lib/global-config');
