+++ src/cli/commands/public-cmds/move-cmd.ts
@@ -1,5 +1,5 @@
 import chalk from 'chalk';
-import { LegacyCommand, CommandOptions } from '../../command';
+import { LegacyCommand, CommandOptions } from '../../legacy-command';
 import { move } from '../../../api/consumer';
 import { PathChangeResult } from '../../../consumer/bit-map/bit-map';
 import { BASE_DOCS_DOMAIN } from '../../../constants';
