+++ src/cli/commands/public-cmds/remote-cmd.ts
@@ -2,7 +2,7 @@
 
 import chalk from 'chalk';
 import Table from 'tty-table';
-import { LegacyCommand, CommandOptions } from '../../command';
+import { LegacyCommand, CommandOptions } from '../../legacy-command';
 import { remoteList, remoteAdd, remoteRm } from '../../../api/consumer';
 import { forEach, empty } from '../../../utils';
 import RemoteUndefined from '../exceptions/remote-undefined';
