+++ src/cli/commands/public-cmds/undeprecate-cmd.ts
@@ -1,7 +1,7 @@
 import R from 'ramda';
 import chalk from 'chalk';
 import { undeprecate } from '../../../api/consumer';
-import { LegacyCommand, CommandOptions } from '../../command';
+import { LegacyCommand, CommandOptions } from '../../legacy-command';
 import { DeprecationResult } from '../../../scope/component-ops/components-deprecation';
 
 export default class Undeprecate implements LegacyCommand {
