+++ src/cli/commands/public-cmds/untrack-cmd.ts
@@ -1,6 +1,6 @@
 import chalk from 'chalk';
 import R from 'ramda';
-import { LegacyCommand, CommandOptions } from '../../command';
+import { LegacyCommand, CommandOptions } from '../../legacy-command';
 import { untrack } from '../../../api/consumer';
 import GeneralError from '../../../error/general-error';
 import { BASE_DOCS_DOMAIN, WILDCARD_HELP } from '../../../constants';
